refector: main 라우팅 정상화
